--- Fixed compilation errors for demo drivers when MPI is not     enabled.
--- Elimination of warnings reported by unused variables
    associated to unstructured meshes when MPI is not enabled

--- Removed unwanted files in the repository from validata folder

    demo_drivers/mpi/distribution/two_d_unstructured_adaptive_poisson/validata



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@1136 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ demo_drivers/mpi/distribution/two_d_unstructured_adaptive_poisson/two_d_parallel_unstructured_adaptive_poisson.cc
@@ -92,7 +92,7 @@ namespace TestArguments
 {
   unsigned Domain_configuration = 1; // Choose the domain geometry
   double Element_size = 1.0e-2; // same as default
-  double Max_adapt = 3;
+  unsigned Max_adapt = 3;
   double Max_permitted_error = 1.0e-3; // same as default
   double Min_permitted_error = 1.0e-5; // same as default
   double Max_element_size = Element_size;
@@ -1479,6 +1479,8 @@ int main(int argc, char* argv[])
  
  problem.mesh_pt()->output(file_initial_distributed_mesh, 2);
 #endif // OOMPH_HAS_MPI
+ 
+ 
  
  // Solve the problem doing (parallel unstructured) mesh adaptation
  problem.newton_solve(TestArguments::Max_adapt);
